Updated config types

+++ src/benchmark/benchmark_config.hpp
@@ -13,17 +13,17 @@
 
 /// @brief Configuration for a benchmark run
 struct BenchmarkConfig {
-  uint64_t data_size;                             ///< Total size of original data
-  uint64_t block_size;                            ///< Size of each block
-  uint64_t num_lost_blocks;                       ///< Number of total blocks lost (recovery + original)
+  size_t data_size;                             ///< Total size of original data
+  size_t block_size;                            ///< Size of each block
+  size_t num_lost_blocks;                       ///< Number of total blocks lost (recovery + original)
   double redundancy_ratio;                        ///< Recovery blocks / original blocks ratio
   int num_iterations;                             ///< Number of iterations to run the benchmark
   uint8_t plot_id;                                ///< Identifier for plotting
   const std::vector<uint32_t>& lost_block_idxs;   ///< Pointer to the lost block indices array
 
   struct {
-    uint32_t num_original_blocks;                 ///< Number of original data blocks
-    uint32_t num_recovery_blocks;                 ///< Number of recovery blocks
+    size_t num_original_blocks;                 ///< Number of original data blocks
+    size_t num_recovery_blocks;                 ///< Number of recovery blocks
   } computed;
 
   bool is_xorec_config;                           ///< Flag to indicate whether this configuration is for XOR-EC algorithm(s)
@@ -32,7 +32,7 @@ struct BenchmarkConfig {
     XorecVersion version;                         ///< Version of the XOR-EC algorithm
 
     bool prefetch;                                ///< Flag to indicate prefetching
-    uint32_t prefetch_bytes;                      ///< Number of bytes to prefetch
+    size_t prefetch_bytes;                      ///< Number of bytes to prefetch
 
     bool gpu_mem;                                 ///< Flag to indicate GPU memory a llocation
     bool touch_gpu_mem;                           ///< Flag to indicate GPU memory warmup
@@ -45,14 +45,14 @@ struct BenchmarkConfig {
 using BenchmarkFunction = void(*)(benchmark::State&, const BenchmarkConfig&);
 
 /// Constants for benchmark configurations (when running full benchmark)
-constexpr uint32_t FIXED_NUM_ORIGINAL_BLOCKS = 128;
-constexpr uint32_t FIXED_NUM_RECOVERY_BLOCKS = 4;
-constexpr uint64_t FIXED_BUFFER_SIZE = 1048576; // 1 MiB
+constexpr size_t FIXED_NUM_ORIGINAL_BLOCKS = 128;
+constexpr size_t FIXED_NUM_RECOVERY_BLOCKS = 4;
+constexpr size_t FIXED_BUFFER_SIZE = 1048576; // 1 MiB
 constexpr double FIXED_PARITY_RATIO = 0.03125;
-constexpr uint64_t FIXED_NUM_LOST_BLOCKS = 1;
-const std::vector<uint64_t> VAR_BUFFER_SIZE = { 134217728, 67108864, 33554432, 16777216, 8388608, 4194304, 2097152, 1048576, 524288, 262144 };
-const std::vector<uint32_t> VAR_NUM_RECOVERY_BLOCKS = { 128, 64, 32, 16, 8, 4, 2, 1 };
-const std::vector<uint32_t> VAR_NUM_LOST_BLOCKS = { 128, 64, 32, 16, 8, 4, 2, 1 };
-const std::vector<uint32_t> VAR_NUM_PREFETCH_BYTES = { 262144, 524288, 1048576, 2097152,};
+constexpr size_t FIXED_NUM_LOST_BLOCKS = 1;
+const std::vector<size_t> VAR_BUFFER_SIZE = { 134217728, 67108864, 33554432, 16777216, 8388608, 4194304, 2097152, 1048576, 524288, 262144 };
+const std::vector<size_t> VAR_NUM_RECOVERY_BLOCKS = { 128, 64, 32, 16, 8, 4, 2, 1 };
+const std::vector<size_t> VAR_NUM_LOST_BLOCKS = { 128, 64, 32, 16, 8, 4, 2, 1 };
+const std::vector<size_t> VAR_NUM_PREFETCH_BYTES = { 262144, 524288, 1048576, 2097152,};
 
 #endif // BENCHMARK_CONFIG_HPP
